Add generator

+++ sources/users/User.cpp
@@ -6,15 +6,15 @@ bool User::isTime(double time) {
     return ((*this).nextJobTime-time>0);
 }
 
-int User::getBudget() {
+double User::getBudget() {
     return (*this).budget;
 }
 
-int User::getSpendings() {
+double User::getSpendings() {
     return (*this).spendings;
 }
 
-bool User::spend(int newSpendings) {
+bool User::spend(double newSpendings) {
     if ((*this).spendings+newSpendings < (*this).budget){
         (*this).spendings += newSpendings;
         return true;
